FIx phần xóa nhân viên, đồ ăn, nguyên liệu

+++ GUI/UserControls_ChildForm/ucDanhSachDoAn.cs
@@ -244,16 +244,22 @@ namespace GUI.UserControls
 
             int[] selectedRows = grvDSDA.GetSelectedRows();
 
-            DialogResult result = XtraMessageBox.Show($"bạn muốn xóa {selectedRows.Length} dòng đã chọn?", "Xóa dữ liệu", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-            if (result == DialogResult.Yes)
+            DialogResult warningDialog = XtraMessageBox.Show("Xóa những dòng này sẽ xóa hết những dòng liên quan ở bảng hóa đơn?\n" +
+                "Bạn vẫn muốn xóa?", "Cảnh báo", MessageBoxButtons.YesNo,MessageBoxIcon.Warning);
+            if(warningDialog == DialogResult.Yes)
             {
-                for (int i = selectedRows.Length - 1; i >= 0; i--)
+                DialogResult result = XtraMessageBox.Show($"Bạn vẫn muốn xóa {selectedRows.Length} dòng đã chọn?", "Xóa dữ liệu", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                if (result == DialogResult.Yes)
                 {
-                    deleteData.Rows.Add(grvDSDA.GetDataRow(selectedRows[i]).ItemArray);
-                    grvDSDA.DeleteRow(selectedRows[i]);
+                    for (int i = selectedRows.Length - 1; i >= 0; i--)
+                    {
+                        deleteData.Rows.Add(grvDSDA.GetDataRow(selectedRows[i]).ItemArray);
+                        grvDSDA.DeleteRow(selectedRows[i]);
+                    }
+                    DeleteData(deleteData);
                 }
-                DeleteData(deleteData);
             }
+
         }
     }
 }
